moved files out of main
- added websocket bridge for browser testing / mcp
