+++ src-tauri/Cargo.toml
@@ -35,6 +35,9 @@ serde_yaml = "0.9"
 reqwest = { version = "0.12", features = ["blocking"] }
 chrono = "0.4"
 tempfile = "3"
+tokio-tungstenite = "0.24"
+futures-util = "0.3"
+tokio = { version = "1", features = ["full"] }
 
 [target.'cfg(not(any(target_os = "android", target_os = "ios")))'.dependencies]
 tauri-plugin-updater = "2"
